requirements.txt
mcstatus
beautifulsoup4
requests
redis
aiohttp[speedups]
colorama
pydantic
python-dotenv
